+++ lib/src/common/config.dart
@@ -1,8 +1,8 @@
 /// The version of the Klutter Pub Plugin.
-const klutterPubVersion = "0.3.0";
+const klutterPubVersion = "1.0.0";
 
 /// The version of the Klutter Gradle Plugin.
-const klutterGradleVersion = "2023.1.1.beta";
+const klutterGradleVersion = "2023.2.1.beta";
 
 /// The version of Kotlin to be used.
 const kotlinVersion = "1.8.20";